requirements/base.txt
Django==4.2.4
psycopg2==2.9.7
django-environ==0.10.0
PyYAML==6.0.1
ua-parser==0.18.0
user-agents==2.2.0
django-user-agents==0.4.0